requirements.txt
Django==5.1.6
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
whitenoise==6.9.0
black==25.1.0
ruff==0.11.0
